requirements.txt
PyYAML>=6.0
fastapi>=0.100
uvicorn>=0.22
python-multipart>=0.0.6

# Optional: for desktop window launcher (scripts/launch_ui_window.py)
# pywebview>=4.0
